Fixed --db-info to work against most old databases * Changed sysnode identifier to a 32-bit integer for backward compatability * Changed start-up logic to allow reading sysnode from old databases * Changed hourly, daily and monthly node identifiers to 32-bit integers * Changed start-up/clean-up logic not to save history if it's not modified

+++ totals.cpp
@@ -13,7 +13,7 @@
 #include "serialize.h"
 #include "preserve.h"
 
-totals_t::totals_t(void) : keynode_t<uint64_t>(1)
+totals_t::totals_t(void) : keynode_t<uint32_t>(1)
 {
    a_hitptime = 0.0;							       /* average hit processing time */
    m_hitptime = 0.0;							       /* maximum hit processing time */
